% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offense.R
\name{count_offense}
\alias{count_offense}
\title{Count NIBRS offense data}
\usage{
count_offense(offense, level = "national", level_detail = NULL)
}
\arguments{
\item{offense}{a string or a list, the offense type(s) to request. Required.}

\item{level}{a string, the level of aggregation for counting. It can be \code{national}, \code{agencies}, \code{regions}, or \code{states}.
If omitted, \code{national} is assumed.}

\item{level_detail}{a string or a list. Use this parameter to indicate what are the agencies/regions/states you want to query.
For example, when \code{level = 'agencies'}, then specify the ORI code of the agencies here.
If \code{level = 'regions'} , then specify what regions to include here. There are four region categories available for choosing: Northeast, Midwest, West, and South.
If \code{level = 'states'}, then specify state abbreviations here, e.g. \code{c('AL', 'AZ')}.
Note, when \code{level = 'national'}, this parameter should be \code{NULL}.}
}
\description{
count_offense() is a function that queries the count of different types of offense from FBI Crime API and returns a data frame based on the query result.
You can specify offense type, level of aggregation (national, agencies, regions, or states), and selected levels (i.e. what are the states/regions/agencies you want to include) in this function.
}
\details{
Before calling any querying functions, you should pass your API KEY with \code{set_fbi_crime_api_key()}.
}
\examples{
count_offense(offense = 'burglary')
count_offense(offense = 'larceny', level = 'agencies', level_detail = 'MA0010100')
count_offense(offense = c('burglary','larceny'), level = 'regions', level_detail = c('South','Northeast'))
count_offense(offense = 'homicide', level = 'states', level_detail = c('AL', 'AZ'))
}
