% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbi_crime_api_key.R
\name{set_fbi_crime_api_key}
\alias{set_fbi_crime_api_key}
\title{Set KEY for FBI Crime API}
\usage{
set_fbi_crime_api_key(x)
}
\arguments{
\item{x}{a string, your key for FBI Crime API.}
}
\value{
set_fbi_crime_api_key() invisibly returns `NULL`.
}
\description{
A convenience function to set key for FBI Crime API as a global option.
Visit the following URL to get a key for the API: \url{https://api.data.gov/signup/}
}
\details{
Before calling any querying functions, you should pass your API KEY with \code{set_fbi_crime_api_key()}.
This function sets the input API key as a global option so that you don't need to specify the key every time.
}
\examples{
set_zillow_web_service_id('SomeExampleKey5AynpvPpUQTEyxE77jo1RU8PIv')

}
