% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrest.R
\name{summarize_arrest}
\alias{summarize_arrest}
\title{Summarize arrest data}
\usage{
summarize_arrest(
  by_offense_type = FALSE,
  offense = NULL,
  level = "national",
  level_detail = NULL,
  variable,
  since,
  until
)
}
\arguments{
\item{by_offense_type}{a boolean. \code{TRUE} means to summarize by offense type, \code{FALSE} means to return information about all offense types.
If omitted, \code{FALSE} is assumed.}

\item{offense}{a string or a list, the offense type(s) to request.
If omitted, \code{NULL} is assumed.}

\item{level}{a string, the level of aggregation. It can be \code{national}, \code{agencies}, \code{regions}, or \code{states}.
If omitted, \code{national} is assumed.}

\item{level_detail}{a string or a list. Use this parameter to indicate what are the agencies/regions/states you want to query.
For example, when \code{level = 'agencies'}, then specify the ORI code of the agencies here.
If \code{level = 'regions'} , then specify what regions to include here. There are four region categories available for choosing: Northeast, Midwest, West, and South.
If \code{level = 'states'}, then specify state abbreviations here, e.g. \code{c('AL', 'AZ')}.
Note, when \code{level = 'national'}, this parameter should be \code{NULL}.}

\item{variable}{a string, the rule for aggregating.
The availale arguments for this parameter vary for different endpoints of the API. Please check the API document (https://crime-data-explorer.fr.cloud.gov/api) for details.
Required.}

\item{since}{a number, the year to start}

\item{until}{a number, the year to end}
}
\description{
summarize_arrest() is a function that queries arrest information from FBI Crime API and returns a data frame.
    You can choose whether to summarize by offense type or not with \code{by_offense_type} and input the offense(s) with \code{offense}.
    You can also specify aggregation level (national, agencies, regions, or states), aggregation variable (age, ethnicity, etc.), and time span in this function.
}
\details{
Before calling any querying functions, you should pass your API KEY with \code{set_fbi_crime_api_key()}.
}
\examples{
summarize_arrest(by_offense_type = TRUE, offense = c('aggravated-assault','rape'), level = 'states', level_detail = c('NY','MA'), variable = 'race', since = 2010, until = 2011)
summarize_arrest(by_offense_type = TRUE, offense = c('aggravated-assault','rape'), variable = 'race', since = 2010, until = 2011)
summarize_arrest(by_offense_type = FALSE, level = 'regions', level_detail = c('South','Midwest'), variable = 'all', since = 2015, until = 2016)
summarize_arrest(by_offense_type = FALSE, variable = 'all', since = 2015, until = 2016)
}
