% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victim.R
\name{summarize_victim}
\alias{summarize_victim}
\title{Summarize NIBRS victim demographic data}
\usage{
summarize_victim(
  offense,
  level = "national",
  level_detail = NULL,
  variable = "count"
)
}
\arguments{
\item{offense}{a string or a list, the offense type(s) to request. Required.}

\item{level}{a string, the level of aggregation. It can be \code{national}, \code{agencies}, \code{regions}, or \code{states}.
If omitted, \code{national} is assumed.}

\item{level_detail}{a string or a list. Use this parameter to indicate what are the agencies/regions/states you want to query.
For example, when \code{level = 'agencies'}, then specify the ORI code of the agencies here.
If \code{level = 'regions'} , then specify what regions to include here. There are four region categories available for choosing: Northeast, Midwest, West, and South.
If \code{level = 'states'}, then specify state abbreviations here, e.g. \code{c('AL', 'AZ')}.
Note, when \code{level = 'national'}, this parameter should be \code{NULL}.}

\item{variable}{a string, the rule for aggregating. It can be \code{age}, \code{ethnicity}, \code{race}, \code{sex}, \code{relationship}, or \code{count}.
If omitted, \code{count} is assumed.}
}
\description{
summarize_victim() is a function that queries victim demographic information from FBI Crime API and returns a data frame.
    You can specify offense type, aggregation level (national, agencies, regions, or states), and aggregation variable (age, ethnicity, etc.) in this function.
}
\details{
Before calling any querying functions, you should pass your API KEY with \code{set_fbi_crime_api_key()}.
}
\examples{
summarize_victim(offense = 'violent-crime')
summarize_victim(offense = 'aggravated-assault', level = 'agencies', level_detail = 'MA0010100', variable = 'age')
summarize_victim(offense = 'property-crime', level = 'regions', level_detail = c('Midwest', 'West'), variable = 'relationship')
summarize_victim(offense = 'motor-vehicle-theft', level = 'states', level_detail = 'DC', variable = 'sex')
summarize_victim(offense = c('violent-crime','property-crime'), level = 'states', level_detail = c('DC','MA'), variable = 'race')
}
