% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg.R
\name{valid_arg_level}
\alias{valid_arg_level}
\title{Validate arguments}
\usage{
valid_arg_level(level, level_detail)
}
\description{
valid_arg_level() is a function used to validate the arguments for \code{level} and \code{level_detail}.
}
\details{
This function is intended to be called by other querying functions within the fbicrime package, but not for external call by users.

It checks two things:
1) whether the input \code{level} is legal (i.e. one of 'agencies', 'national', 'regions', 'states');
2) whether the input \code{level} matches \code{level_detail}.
This function produces an error when \code{level} is 'national' but \code{level_detail} is not NULL,
and when \code{level} is not 'national' but \code{level_detail} is NULL or is not specified.
}
